migrated some parts of api logic to library
