+++ spliff-lib/Cargo.toml
@@ -1,12 +1,11 @@
 [package]
-name = "spliff-api"
+name = "spliff-lib"
 version = "0.1.0"
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rocket = { version = "0.5.0-rc.1", features = ["json"]}
 solana-client = "1.7.9"
 solana-sdk = "1.7.9"
 spl-token = "3.2.0"
@@ -14,6 +13,3 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.59"
 spl-associated-token-account = "1.0.3"
 solana-account-decoder = "1.7.7"
-
-#spl-lib
-spliff-lib = {version = "0.1.0", path="../spliff-lib" }
