+++ spliff-lib/src/lib.rs
@@ -1,0 +1,14 @@
+pub mod account;
+pub mod state;
+pub mod token;
+pub fn hello() -> String {
+    "Hello".to_string()
+}
+
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn it_works() {
+        assert_eq!(2 + 2, 4);
+    }
+}
